requirements.txt
vosk
sounddevice
numpy
thefuzz
pyttsx3
colorama
customtkinter
Pillow
keyboard
pystray